+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Version History
 
+## 1.1.0 / 2021-08-09
+
+- Added stored value retrieval
+- Added URLencode() wrapper to data view URL
+
 ## 1.0.2 / 2021-07-13
 
 - Update sampleoutput_placeholder.csv file